This is actually the previous commit (WTF)

+++ eve-site/src/main/java/nz/net/dnh/eve/business/AbstractType.java
@@ -33,5 +33,5 @@ public interface AbstractType {
 	/**
 	 * @return The id of the type
 	 */
-	public long getId();
+	public int getId();
 }
